add an asset abstract base class
