+++ starfish/asset/__init__.py
@@ -1,5 +1,6 @@
 
 
-from starfish.asset.asset import Asset
+from starfish.asset.asset_base import AssetBase
 from starfish.asset.memory_asset import MemoryAsset
 from starfish.asset.squid_asset import SquidAsset
+from starfish.asset.asset import Asset
